[feat]: 플레이어 관련 기능 추가 및 스폰 타이밍 개선, 인게임 매니저 데이터 수정

+++ Assets/00_WorkSpace/YSJ/Scripts/PhotonNetworkCustomProperties.cs
@@ -775,7 +775,7 @@ public static class PhotonNetworkCustomProperties
         );
     }
 
-    public static void LocalPlayerRaceFinishedSetting(float finishTime)
+    public static void LocalPlayerRaceFinishedSetting(double finishTime)
     {
         if (!PhotonNetwork.InRoom) return;
